Refactored simplespans to a separate package
